Refactoring the app using Stack Navigator
