+++ app/components/UserProfile.js
@@ -1,0 +1,20 @@
+import { View, StyleSheet, Text } from 'react-native';
+import React from 'react';
+
+const UserProfile = () => {
+  return (
+    <View style={styles.container}>
+      <Text>Profile</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default UserProfile;
