+++ package.json
@@ -9,6 +9,8 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/native-stack": "^6.9.12",
     "axios": "^1.3.5",
     "expo": "~48.0.10",
     "expo-image-picker": "~14.1.1",
